move dataset test utilities under arrow/dataset/
