+++ cpp/src/arrow/testing/test_data.h.in
@@ -33,41 +33,4 @@ std::string ArrowTestDataPath() { return ARROW_TEST_DATA_PATH; }
 
 std::string ParquetTestDataPath() { return PARQUET_TEST_DATA_PATH; }
 
-// Convenience class allowing to retrieve integration data easily, e.g.
-//
-// class MyTestClass : public ::testing::Test, TestDataFixtureMixin {
-// }
-//
-// TEST(MyTestClass, ATest) {
-//    this->OpenParquetFile("data/binary.parquet")
-//    // Do something
-// }
-class TestDataFixtureMixin : public ::testing::Test {
- public:
-  TestDataFixtureMixin() {
-    auto root = std::make_shared<fs::LocalFileSystem>();
-    arrow_fs_ = std::make_shared<fs::SubTreeFileSystem>(ArrowTestDataPath(), root);
-    parquet_fs_ = std::make_shared<fs::SubTreeFileSystem>(ParquetTestDataPath(), root);
-  }
-
-  std::shared_ptr<io::RandomAccessFile> OpenParquetFile(const std::string& relative_path) {
-    std::shared_ptr<io::RandomAccessFile> file;
-    ARROW_EXPECT_OK(ParquetDataFileSystem()->OpenInputFile(relative_path, &file));
-    return file;
-  }
-
-  std::shared_ptr<io::RandomAccessFile> OpenArrowFile(const std::string& relative_path) {
-    std::shared_ptr<io::RandomAccessFile> file;
-    ARROW_EXPECT_OK(ArrowDataFileSystem()->OpenInputFile(relative_path, &file));
-    return file;
-  }
-
-  fs::FileSystem* ArrowDataFileSystem() const { return arrow_fs_.get(); }
-  fs::FileSystem* ParquetDataFileSystem() const { return parquet_fs_.get(); }
-
- protected:
-  std::shared_ptr<fs::FileSystem> arrow_fs_;
-  std::shared_ptr<fs::FileSystem> parquet_fs_;
-};
-
 }  // namespace arrow
