--- cpp/src/parquet/arrow/io_test.h
@@ -1,67 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-#pragma once
-
-#include <gtest/gtest.h>
-#include <memory>
-
-#include "parquet/arrow/writer.h"
-
-#include "arrow/buffer.h"
-#include "arrow/record_batch.h"
-#include "arrow/table.h"
-#include "arrow/testing/gtest_util.h"
-
-using ::arrow::Buffer;
-using ::arrow::RecordBatch;
-using ::arrow::RecordBatchReader;
-using ::arrow::Status;
-using ::arrow::Table;
-
-namespace parquet {
-namespace arrow {
-
-class ArrowParquetWriterMixin : public ::testing::Test {
- public:
-  std::shared_ptr<Buffer> Write(RecordBatchReader* reader) {
-    auto pool = ::arrow::default_memory_pool();
-
-    std::shared_ptr<Buffer> out;
-    auto sink = CreateOutputStream(pool);
-
-    ARROW_EXPECT_OK(WriteRecordBatchReader(reader, pool, sink));
-    ARROW_EXPECT_OK(sink->Finish(&out));
-
-    return out;
-  }
-
-  std::shared_ptr<Buffer> Write(const Table& table) {
-    auto pool = ::arrow::default_memory_pool();
-
-    std::shared_ptr<Buffer> out;
-    auto sink = CreateOutputStream(pool);
-
-    ARROW_EXPECT_OK(WriteTable(table, pool, sink, 1U << 16));
-    ARROW_EXPECT_OK(sink->Finish(&out));
-
-    return out;
-  }
-};
-
-}  // namespace arrow
-}  // namespace parquet
